removed caching; renamed app_settings to settings; removed extra call to get user_agent
